refactor(*): initial working refactor
