--- app/auth/token.go
@@ -1,104 +1,0 @@
-package auth
-
-import (
-	"encoding/json"
-	"fmt"
-	"log"
-	"net/http"
-	"os"
-	"strconv"
-	"strings"
-	"time"
-
-	jwt "github.com/dgrijalva/jwt-go"
-)
-
-// CreateToken : Creates a jwt token for a session
-func CreateToken(userID uint32) (string, error) {
-	claims := jwt.MapClaims{}
-
-	claims["authorized"] = true
-	claims["user_id"] = userID
-	claims["exp"] = time.Now().Add(time.Hour * 168).Unix() //Token expires after 168 hours, 7 days (1 week)
-
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-
-	return token.SignedString([]byte(os.Getenv("API_SECRET")))
-
-}
-
-// IsTokenValid : Checks if the token if valid
-func IsTokenValid(r *http.Request) error {
-	tokenString := ExtractToken(r)
-
-	token, err := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
-		if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
-			return nil, fmt.Errorf("Unexpected signing method: %v", token.Header["alg"])
-		}
-		return []byte(os.Getenv("API_SECRET")), nil
-	})
-
-	if err != nil {
-		return err
-	}
-
-	if claims, ok := token.Claims.(jwt.MapClaims); ok && token.Valid {
-		Pretty(claims)
-	}
-
-	return nil
-}
-
-// ExtractToken : Extracts the token from the request headers or the url
-func ExtractToken(r *http.Request) string {
-	keys := r.URL.Query()
-	token := keys.Get("access_token")
-
-	if token != "" {
-		return token
-	}
-
-	bearerToken := r.Header.Get("Authorization")
-	if len(strings.Split(bearerToken, " ")) == 2 {
-		return strings.Split(bearerToken, " ")[1]
-	}
-
-	return ""
-}
-
-// ExtractTokenID : Extracts the "user_id" from the token
-func ExtractTokenID(r *http.Request) (uint, error) {
-
-	tokenString := ExtractToken(r)
-
-	token, err := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
-		if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
-			return nil, fmt.Errorf("Unexpected signing method: %v", token.Header["alg"])
-		}
-		return []byte(os.Getenv("API_SECRET")), nil
-	})
-
-	if err != nil {
-		return 0, err
-	}
-
-	claims, ok := token.Claims.(jwt.MapClaims)
-	if ok && token.Valid {
-		uid, err := strconv.ParseUint(fmt.Sprintf("%.0f", claims["user_id"]), 10, 32)
-		if err != nil {
-			return 0, err
-		}
-		return uint(uid), nil
-	}
-
-	return 0, nil
-}
-
-// Pretty : Displays the claims licely in the terminal
-func Pretty(data interface{}) {
-	_, err := json.MarshalIndent(data, "", " ")
-	if err != nil {
-		log.Println(err)
-		return
-	}
-}
