--- app/responses/json.go
@@ -1,31 +1,0 @@
-package responses
-
-import (
-	"encoding/json"
-	"fmt"
-	"net/http"
-)
-
-// JSON : Response with Json data
-func JSON(w http.ResponseWriter, statusCode int, data interface{}) {
-	w.WriteHeader(statusCode)
-
-	err := json.NewEncoder(w).Encode(data)
-	if err != nil {
-		fmt.Fprintf(w, "%s", err.Error())
-	}
-}
-
-// ERROR : Error response with Json data
-func ERROR(w http.ResponseWriter, statusCode int, err error) {
-	if err != nil {
-		JSON(w, statusCode, struct {
-			Error string `json:"error"`
-		}{
-			Error: err.Error(),
-		})
-		return
-	}
-
-	JSON(w, http.StatusBadRequest, nil)
-}
