+++ cmd/api/main.go
@@ -3,16 +3,15 @@ package api
 import (
 	"flag"
 	"fmt"
+	"github.com/Victoria-engine/api-v2/pkg/api"
 	"log"
 	"os"
 
-	"github.com/Victoria-engine/api-v2/app/services"
 	"github.com/joho/godotenv"
 )
 
-var server = services.Server{}
 
-// Run : Runs the REST API
+// Run : Runs the REST API and setups flags
 func Run() {
 	var (
 		port = flag.String("port", os.Getenv("PORT"), "The server port")
@@ -28,17 +27,17 @@ func Run() {
 		fmt.Println("Got the env values !")
 	}
 
-	server.Initialize(
+
+	err = api.Start(
 		os.Getenv("DB_DRIVER"),
 		os.Getenv("DB_USER"),
 		os.Getenv("DB_PASSWORD"),
 		os.Getenv("DB_PORT"),
 		os.Getenv("DB_HOST"),
 		os.Getenv("DB_NAME"),
-	)
-
-	// Seed in case you need dummy data
-	//seed.Load(server.DB)
-
-	server.Run(fmt.Sprintf(":%s", *port))
+		*port,
+		)
+	if err != nil {
+		panic(err.Error())
+	}
 }
