+++ pkg/services/users_listener.go
@@ -1,0 +1,168 @@
+package services
+
+import (
+	"errors"
+	"fmt"
+	"github.com/Victoria-engine/api-v2/pkg/utl/jwtauth"
+	"net/http"
+	"strconv"
+
+	"github.com/Victoria-engine/api-v2/pkg/models"
+	"github.com/Victoria-engine/api-v2/pkg/presenters"
+	"github.com/Victoria-engine/api-v2/pkg/utl/responses"
+	"github.com/gorilla/mux"
+)
+
+// // CreateUser : Creates a new user
+// func (server *Server) CreateUser(w http.ResponseWriter, r *http.Request) {
+// 	if r.Method != http.MethodPost {
+// 		w.WriteHeader(405) // Method not allowed
+// 		return
+// 	}
+
+// 	body, err := ioutil.ReadAll(r.Body)
+// 	if err != nil {
+// 		responses.ERROR(w, http.StatusUnprocessableEntity, err)
+// 	}
+
+// 	user := models.User{}
+
+// 	err = json.Unmarshal(body, &user)
+// 	if err != nil {
+// 		responses.ERROR(w, http.StatusUnprocessableEntity, err)
+// 		return
+// 	}
+
+// 	user.Prepare()
+
+// 	err = user.Validate("")
+// 	if err != nil {
+// 		responses.ERROR(w, http.StatusUnprocessableEntity, err)
+// 		return
+// 	}
+
+// 	userCreated, err := user.SaveUser(server.DB)
+
+// 	if err != nil {
+// 		formattedError := formaterror.FormatError(err.Error())
+// 		responses.ERROR(w, http.StatusInternalServerError, formattedError)
+// 		return
+// 	}
+
+// 	w.Header().Set("Location", fmt.Sprintf("%s%s/%d", r.Host, r.RequestURI, userCreated.ID))
+// 	responses.JSON(w, http.StatusCreated, userCreated)
+// }
+
+// GetUserInfo : Gets the user information
+func (server *Server) GetUserInfo(w http.ResponseWriter, r *http.Request) {
+	vars := mux.Vars(r)
+
+	uid, err := strconv.ParseUint(vars["id"], 10, 32)
+	if err != nil {
+		responses.ERROR(w, http.StatusBadRequest, err)
+		return
+	}
+
+	user := models.User{}
+
+	foundUser, err := user.FindUserByID(server.DB, uint(uid))
+	if err != nil {
+		responses.ERROR(w, http.StatusNotFound, err)
+		return
+	}
+
+	res := presenters.GetUserInfoPresenter(foundUser)
+
+	responses.JSON(w, http.StatusOK, res)
+}
+
+// // UpdateUser : Updates a user
+// func (server *Server) UpdateUser(w http.ResponseWriter, r *http.Request) {
+// 	if r.Method != http.MethodPut {
+// 		w.WriteHeader(405) // Method not allowed
+// 		return
+// 	}
+
+// 	vars := mux.Vars(r)
+// 	uid, err := strconv.ParseUint(vars["id"], 10, 32)
+// 	if err != nil {
+// 		responses.ERROR(w, http.StatusBadRequest, err)
+// 		return
+// 	}
+
+// 	body, err := ioutil.ReadAll(r.Body)
+// 	if err != nil {
+// 		responses.ERROR(w, http.StatusUnprocessableEntity, err)
+// 		return
+// 	}
+
+// 	user := models.User{}
+
+// 	err = json.Unmarshal(body, &user)
+// 	if err != nil {
+// 		responses.ERROR(w, http.StatusUnprocessableEntity, err)
+// 		return
+// 	}
+
+// 	tokenID, err := authmiddleware.ExtractTokenID(r)
+// 	if err != nil {
+// 		responses.ERROR(w, http.StatusUnauthorized, errors.New("Unauthorized"))
+// 		return
+// 	}
+
+// 	if tokenID != uint32(uid) {
+// 		responses.ERROR(w, http.StatusUnauthorized, errors.New(http.StatusText(http.StatusUnauthorized)))
+// 		return
+// 	}
+
+// 	user.Prepare()
+
+// 	err = user.Validate("update")
+// 	if err != nil {
+// 		responses.ERROR(w, http.StatusUnprocessableEntity, err)
+// 		return
+// 	}
+
+// 	updatedUser, err := user.UpdateUser(server.DB, uint32(uid))
+// 	if err != nil {
+// 		formattedError := formaterror.FormatError(err.Error())
+// 		responses.ERROR(w, http.StatusInternalServerError, formattedError)
+// 		return
+// 	}
+
+// 	responses.JSON(w, http.StatusOK, updatedUser)
+// }
+
+// DeleteUser : Deletes a user
+func (server *Server) DeleteUser(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodDelete {
+		w.WriteHeader(405) // Method not allowed
+		return
+	}
+
+	vars := mux.Vars(r)
+
+	user := models.User{}
+
+	uid, err := strconv.ParseUint(vars["id"], 10, 32)
+	if err != nil {
+		responses.ERROR(w, http.StatusBadRequest, err)
+		return
+	}
+	tokenID, err := jwt.ExtractTokenID(r)
+	if err != nil {
+		responses.ERROR(w, http.StatusUnauthorized, errors.New("Unauthorized"))
+		return
+	}
+	if tokenID != 0 && tokenID != uint(uid) {
+		responses.ERROR(w, http.StatusUnauthorized, errors.New(http.StatusText(http.StatusUnauthorized)))
+		return
+	}
+	_, err = user.DeleteByID(server.DB, uint(uid))
+	if err != nil {
+		responses.ERROR(w, http.StatusInternalServerError, err)
+		return
+	}
+	w.Header().Set("Entity", fmt.Sprintf("%d", uid))
+	responses.JSON(w, http.StatusNoContent, "")
+}
