+++ pkg/utl/crypt/crypt.go
@@ -1,0 +1,13 @@
+package crypt
+
+import "golang.org/x/crypto/bcrypt"
+
+// HashPassword : Hashes the a password
+func HashPassword(password string) ([]byte, error) {
+	return bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
+}
+
+// VerifyPassword : Verifies a password
+func VerifyPassword(hashedPassword, password string) error {
+	return bcrypt.CompareHashAndPassword([]byte(hashedPassword), []byte(password))
+}
