+++ pkg/utl/middleware/json/json.go
@@ -1,0 +1,12 @@
+package json
+
+import "net/http"
+
+// SetMiddlewareJSON : Converts the requests to JSON format
+func SetMiddlewareJSON(next http.HandlerFunc) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "application/json")
+		next(w, r)
+	}
+}
+
